Updated connectors
The IDE44 pin connector was wrong. Due to the complexity of having 3
connectors on board the size of the board need to be increased. Now it
is almost like a Credit Card Size. It needs an Amiga 1200 or Amiga 600
placed into a TowerCase and *NOT TOO NEAR* to the bottom of the case...
But if you need a 4 devices connected, for sure the default case has not
accomplishes that!